# tea-spec/openapi.yaml
...
openapi: 3.1.1
...
info:
  title: Transparency Exchange API
  summary: The OWASP Transparency Exchange API specification for consumers and publishers
  description: TBC
  contact:
    name: TEA Working Group
    email: tbc@somewhere.tld
    url: https://github.com/CycloneDX/transparency-exchange-api
  license:
    name: Apache 2.0
    url: https://github.com/CycloneDX/transparency-exchange-api/blob/main/LICENSE
  version: 0.4.0
servers:
  - url: http://localhost/tea/v1
    description: Local development
paths:
...
  /discovery:
    get:
      description: Discovery endpoint which resolves TEI into product release UUID.
      operationId: discoveryByTei
      parameters:
        - name: tei
          in: query
          required: true
          description: Transparency Exchange Identifier (TEI) for the product being discovered. Provide the TEI as a URL-encoded string per RFC 3986, RFC 3987.
          schema:
            type: string
            example: urn%3Atei%3Auuid%3Aproducts.example.com%3Ad4d9f54a-abcf-11ee-ac79-1a52914d44b
      responses:
        '200':
          $ref: "#/components/responses/discovery-response"
        '400':
          $ref: "#/components/responses/400-invalid-request"
        '404':
          $ref: "#/components/responses/404-object-by-id-not-found"
      tags:
        - TEA Discovery
components:
  schemas:
...
    uuid:
      type: string
      description: A UUID
      format: uuid
      pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"

    #
    # TEA Product
    #
...
    unknown-error-type:
      type: string
      description: Classification of TEA error response
      enum:
        - OBJECT_UNKNOWN
        - OBJECT_NOT_SHAREABLE
    #
    # Types used in API responses
    #
    error-response:
      type: object
      description: Error response
      additionalProperties: false
      properties:
        error:
          $ref: "#/components/schemas/unknown-error-type"
      required:
        - error
    tea-server-info:
      type: object
      description: TEA server information including URL, versions, and optional priority
      additionalProperties: false
      properties:
        rootUrl:
          description: Root URL of the TEA server for this TEI without trailing slash
          type: string
          format: uri
          example: https://api.teaexample.com
        versions:
          description: Supported TEA API versions at this server without v prefix
          type: array
          minItems: 1
          items:
            type: string
          example: ["0.2.0-beta.2", "1.0.0"]
        priority:
          description: Optional priority for this server (0.0 to 1.0, where 1.0 is highest priority)
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
          example: 0.8
      required:
        - rootUrl
        - versions
    discovery-info:
      type: object
      description: Discovery information for a TEI
      additionalProperties: false
      properties:
        productReleaseUuid:
          description: UUID of the resolved TEA Product Release
          $ref: "#/components/schemas/uuid"
          example: d4d9f54a-abcf-11ee-ac79-1a52914d44b
        servers:
          description: Array of TEA server information
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/tea-server-info"
      required:
        - productReleaseUuid
        - servers

    #
    # CLE (Common Lifecycle Enumeration) objects
    # Based on ECMA-428 TC54 TG3 CLE Specification v1.0.0
    #
...
  responses:
...
    400-invalid-request:
      description: Request was Invalid
      content:
        application/json: {}
...
    404-object-by-id-not-found:
      description: Object requested by identifier not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/error-response"
    discovery-response:
      description: Discovery information for the requested TEI
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: "#/components/schemas/discovery-info"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    basicAuth:
      type: http
      scheme: basic
security:
  - bearerAuth: []
  - basicAuth: []
...